+++ Skyress/DTOs/Items/UpdateItemPriceRequest.cs
@@ -1,3 +1,3 @@
 namespace Skyress.API.DTOs.Items;
 
-public record UpdateItemPriceRequest(double Price);
+public record UpdateItemPriceRequest(decimal Price);
